added memcpy and edited  structure of library
