+++ ydb/library/yql/utils/simd/exec/merge_columns/AVX2_algo/avx2_algo.cpp
@@ -1,9 +1,9 @@
-#include <ydb/library/yql/utils/simd/exec/merge_columns/merge.h>
+#include <ydb/library/yql/utils/simd/simd.h>
 #include <immintrin.h>
 #include <avxintrin.h>
 
 template<>
-THolder<Perfomancer::Interface> Perfomancer::Create<AVX2Trait>() {
+THolder<NSimd::Perfomancer::Interface> NSimd::Perfomancer::Create<NSimd::AVX2Trait>() {
     Cerr << "AVXTrait ";
     return MakeHolder<Algo<AVX2Trait>>();
 }
