+++ ydb/library/yql/utils/simd/exec/merge_columns/SSE42_algo/sse42_algo.cpp
@@ -1,8 +1,8 @@
-#include <ydb/library/yql/utils/simd/exec/merge_columns/merge.h>
+#include <ydb/library/yql/utils/simd/simd.h>
 #include <immintrin.h>
 
 template<>
-THolder<Perfomancer::Interface> Perfomancer::Create<SSE42Trait>() {
+THolder<NSimd::Perfomancer::Interface> NSimd::Perfomancer::Create<NSimd::SSE42Trait>() {
     Cerr << "SSETrait ";
     return MakeHolder<Algo<SSE42Trait>>();
 }
